% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqspRex.R
\name{sqspRex}
\alias{sqspRex}
\title{Extract download links from XML}
\usage{
sqspRex(file.types, local.path, site.alias)
}
\arguments{
\item{file.types}{File type suffixes as a vector.}

\item{local.path}{The local path to the XML file.}

\item{site.alias}{The Squarespace alias of your site most likely in the format of a
subdomain such as "my-site.squarespace.com without any "https://" or "http://".}
}
\value{
A dataframe of downloadable links.
}
\description{
Get a dataframe of download links from the standard XML Squarespace export.
}
\details{
The only way to export from Squarespace is a Wordpress format XML file.
If all you are trying to do is recover the downloadable files from your
Squarespace site rather than migrate to Wordpress this will involve lot of
unnecessary time and effort.
This package takes the XML export and produces a dataframe of downloadable
file links.
You can then download the files from the links at your leisure.
}
\examples{
sqspRex(file.types, local.path, site.stem)
sqspRex(c(pdf, doc, docx, xls), "~/Desktop/export.xml", "my-site.squarespace.com")
}
