% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqspRimg.R
\name{sqspRimg}
\alias{sqspRimg}
\title{Extract image download links from XML}
\usage{
sqspRimg(file.types, local.path)
}
\arguments{
\item{file.types}{File type suffixes as a vector.}

\item{local.path}{The local path to the XML file.}
}
\value{
A dataframe of downloadable links.
}
\description{
Get a dataframe of image download links from the standard XML Squarespace export.
}
\details{
The only way to export from Squarespace is a Wordpress format XML file.
If all you are trying to do is recover the images from your
Squarespace site rather than migrate to Wordpress this will involve lot of
unnecessary time and effort.
This package takes the XML export and produces a dataframe of downloadable
links to previously uploaded site images.
You can then download the files from the links at your leisure.
}
\examples{
sqspRimg(file.types, local.path)
sqspRimg(c(jpg, png, bmp), "~/Desktop/export.xml")
}
